+++ D3D11Engine/WorldObjects.cpp
@@ -40,6 +40,7 @@ bool VobInfo::ComputeIndoorLightMask() const {
 void VobInfo::UpdateState() {
     WorldMatrix = *Vob->GetWorldMatrixPtr();
     LastRenderPosition = Vob->GetPositionWorld();
+    LastRenderBBox = Vob->GetBBox();
 
     const bool currentIndoorVob = Vob->IsIndoorVob();
     if ( !HasIndoorLightMaskSample || currentIndoorVob != IsIndoorVob ) {